Here's my commit message for the changes I've made:
fix: Resolve circular dependency between handlers and services

I've refactored your codebase to eliminate a circular import dependency between the 'handlers' and 'services' packages.

- I introduced a new 'dto' package (`trello/dto`).
- I moved all Data Transfer Object (DTO) struct definitions from `trello/handlers/dto.go` to the new `trello/dto` package, organizing them into logical files (e.g., `board_dto.go`, `card_dto.go`).
- I relocated all DTO mapping functions (e.g., `MapBoardToResponse`, `MapCardToResponse`) from their respective handler files (e.g., `trello/handlers/board_handler.go`) into the corresponding files within the new `trello/dto` package.
- I updated import paths in the 'handlers' package to use the DTOs and mapping functions from the new 'dto' package.
- I updated import paths in the 'services' package (including `broadcast_helper.go`) to use DTOs and mapping functions from the new 'dto' package. Crucially, this removes the dependency of the 'services' package on the 'handlers' package for DTO mapping.

This change ensures a cleaner separation of concerns and a more maintainable project structure by breaking the import cycle. Handlers now depend on services (for interfaces) and DTOs, while services depend on DTOs but no longer on handlers.
